Add article translation feature
Adds a method to get article translations and load article translation
table if it is not already downloaded. Methods for sending and receiving
articles from and to the Bring API have been extended to translate
article names if a locale has been provided as argument.
Also adds a method to get user settings, which includes also language
settings for each list.